+++ CHANGES.md
@@ -74,6 +74,7 @@
 ##Fixes
 - since rc.54
   - `Required.value()` is correctly translated to `Validate.message()` (see [#373](https://github.com/klum-dsl/klum-ast/issues/373))
+  - CopyHandler should ignore field type IGNORED (see [#374](https://github.com/klum-dsl/klum-ast/issues/374))
 - since rc.51
   - root objects of the wrong type should silently be ignored. This allows partial models to be created, which is especially
     important for testing.